+++ components/bottombar/api/src/main/java/com/flipperdevices/bottombar/model/BottomBarTab.kt
@@ -1,0 +1,7 @@
+package com.flipperdevices.bottombar.model
+
+enum class BottomBarTab {
+    DEVICE,
+    ARCHIVE,
+    HUB
+}