+++ components/faphub/installation/button/impl/src/main/java/com/flipperdevices/faphub/installation/button/impl/api/FapInstallationUIApiImpl.kt
@@ -5,6 +5,8 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
+import com.flipperdevices.bottombar.api.BottomNavigationHandleDeeplink
+import com.flipperdevices.bottombar.model.BottomBarTab
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.core.ui.ktx.placeholderConnecting
 import com.flipperdevices.faphub.installation.button.api.FapButtonConfig
@@ -14,8 +16,10 @@ import com.flipperdevices.faphub.installation.button.impl.composable.ComposableF
 import com.flipperdevices.faphub.installation.button.impl.composable.ComposableFapInstallButton
 import com.flipperdevices.faphub.installation.button.impl.composable.ComposableFapInstalledButton
 import com.flipperdevices.faphub.installation.button.impl.composable.ComposableFapInstallingButton
+import com.flipperdevices.faphub.installation.button.impl.composable.ComposableFapNoInstallButton
 import com.flipperdevices.faphub.installation.button.impl.composable.ComposableFapUpdateButton
 import com.flipperdevices.faphub.installation.button.impl.composable.ComposableFapUpdatingButton
+import com.flipperdevices.faphub.installation.button.impl.composable.ComposableFlipperNotConnectedButton
 import com.flipperdevices.faphub.installation.button.impl.viewmodel.FapStatusViewModel
 import com.flipperdevices.faphub.installation.stateprovider.api.model.FapState
 import com.squareup.anvil.annotations.ContributesBinding
@@ -23,7 +27,9 @@ import tangle.viewmodel.compose.tangleViewModel
 import javax.inject.Inject
 
 @ContributesBinding(AppGraph::class, FapInstallationUIApi::class)
-class FapInstallationUIApiImpl @Inject constructor() : FapInstallationUIApi {
+class FapInstallationUIApiImpl @Inject constructor(
+    private val bottomBarApi: BottomNavigationHandleDeeplink
+) : FapInstallationUIApi {
     @Composable
     override fun ComposableButton(
         config: FapButtonConfig?,
@@ -80,13 +86,18 @@ class FapInstallationUIApiImpl @Inject constructor() : FapInstallationUIApi {
                 modifier = modifier
             )
 
-            FapState.ConnectFlipper -> ComposableFapInstallButton(
+            FapState.ConnectFlipper -> ComposableFlipperNotConnectedButton(
                 modifier = modifier,
                 fapButtonSize = fapButtonSize,
-                onClick = { }
+                onOpenDeviceTab = {
+                    bottomBarApi.onChangeTab(
+                        tab = BottomBarTab.DEVICE,
+                        force = true
+                    )
+                }
             )
 
-            FapState.FlipperOutdated -> ComposableFapInstalledButton(
+            FapState.FlipperOutdated -> ComposableFapNoInstallButton(
                 modifier = modifier,
                 fapButtonSize = fapButtonSize
             )