+++ components/faphub/installation/button/impl/src/main/java/com/flipperdevices/faphub/installation/button/impl/composable/dialogs/ComposableFlipperNotConnectedDialog.kt
@@ -1,0 +1,24 @@
+package com.flipperdevices.faphub.installation.button.impl.composable.dialogs
+
+import androidx.compose.runtime.Composable
+import com.flipperdevices.core.ui.dialog.composable.FlipperDialog
+import com.flipperdevices.faphub.installation.button.impl.R
+import com.flipperdevices.core.ui.res.R as DesignSystem
+
+@Composable
+fun ComposableFlipperNotConnectedDialog(
+    onDismiss: () -> Unit,
+    onOpenDeviceTab: () -> Unit
+) {
+    FlipperDialog(
+        titleId = R.string.faphub_installation_dialog_flipper_not_connected_title,
+        textId = R.string.faphub_installation_dialog_flipper_not_connected_desc,
+        buttonTextId = R.string.faphub_installation_dialog_flipper_not_connected_btn,
+        onClickButton = {
+            onDismiss()
+            onOpenDeviceTab()
+        },
+        imageId = DesignSystem.drawable.ic_flipper_upload_failed,
+        onDismissRequest = onDismiss
+    )
+}