+++ CHANGELOG.md
@@ -4,11 +4,13 @@
 
 - [FIX] Add confirm delete dialog and some small fixes
 - [FIX] Hide self updater dialog on click, add self updater in debug mode, remove nfc mfkey32 from options
+- [FIX] Text alignment for simple device info
 - [Feature] FapHub MVPv0
 - [Feature] Update README with Fdroid badge
 - [Feature] Migrate faphub to target flow
 - [Feature] Update Geminio for create modules - simple, navigation, api/impl
 - [Feature] FapHub manifest support offline mode and add target support for buttons 
+- [Feature] Add flipper not connected dialog
 - [Refactor] Migrate all feature components modules from KSP to Anvil
 - [Refactor] Migrate to Ktorfit
 - [Refactor] Migrate key emulate to new module, decompose Emulate Helper